Filter and Cloned
filter() is useful for interators and seperating out values depending on
a predicate that we create.

cloned() is the exact same as .map(|x| x) over an iterator, this is very
useful and should be remembered when you come across a compiler warning
like so:

'value of type <_> cannot be build from std::iter::Iterator<Item=&<_>>'

it hints that the value you seek is behind a pointer (or two in this
case) that needs to be dereferenced appropriately.

this was a tough one!